requirements.txt
aiogram~=3.4.1
asyncio~=3.4.3
psycopg2-binary
python-dotenv~=1.0.1
watchdog~=4.0.0
PyYAML~=6.0.1